Feat: searchResult component 생성
